# Optimize RelevantFeatureAugmenter to avoid re-extraction