@nils-braun 2020-04-19
Hi @pnb! Thank you very much for this well-documented and prepared PR! Nice you solved that problem.
Your code is fine, however I have a general comment: it involves a lot of code-duplication. Therefore my suggestion:
* create a function, maybe name it `_fit_and_augment` which contains all the code of the `fit` function and returns the `X_augmented`
* call this function in `fit` (and still return `self`) and also in `fit_transform` and add the additional few lines.
How does this sound?