@pnb 2020-04-19
@nils-braun Sounds good! I didn't like the amount of code duplication either, to be honest. I will make that change.